Remove ImageJ 1.x and Micro-Manager artifacts
It was incorrect to include an MMCoreJ.jar but then access the device
adapters (and the native Micro-Manager library) from an installed
Micro-Manager version; It is *highly* dangerous to assume that a random
version of MMCoreJ.jar works with a random Micro-Manager version because
it is *prone* to lead to crashes due to interface mismatches.

So let's not do that. Not at all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>