Introduce a wrapper for accessing CMMCore via an installed MMCoreJ.jar
MMCoreJ.jar versions really bound tightly to the corresponding native
Micro-Manager library; Using a hard-coded MMCoreJ.jar with a native
library from a random Micro-Manager installation is very likely to result
only in crashes or even worse problems.

The solution is to introduce a wrapper that uses reflection to access a
given MMCoreJ.jar.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>