Ensure that the ImageJ2 plugins can be found by KNIME
KNIP accepts only a certain subset of ImageJ2 plugins because it has to
ensure that there are widgets for all the input/output parameters.
Unfortunately, KNIP does not mention it when it discards ImageJ2 plugins
that cannot be handled. But this developer had Christian Dietz sitting
next to him (which does not scale) and could figure out that the ImageJ
common versions of the former ImgLib2 meta classes need to be used.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>