Add a Micro-Manager configurator
Micro-Manager needs to be configured so that it is known where the
Micro-Manager installation lives on this particular computer, and
also to provide a hardware configuration as written out into a file
by Micro-Manager.

Ideally, the configurator would provide one output port that
is then connected to all nodes that require a Micro-Manager
configuration. Unfortunately, this is not possible via ImageJ2
plugins yet; KNIP needs to be extended to allow for input/output
parameters that are provided in the form of ports. As a workaround,
the configurator node needs to be present, and needs to be executed
before any node that wants to access the Micro-Manager instance.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>